product/viewsets/category_viewset.py
from rest_framework.viewsets import ModelViewSet

from product.models import Category
from product.serializers.category_serializer import CategorySerializer


class CategoryViewSet(ModelViewSet):
    serializer_class = CategorySerializer

    def get_queryset(self):
        # para definir a ordenação pelo id
        return Category.objects.all().order_by("id")
