pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = ["maxh33 <contact@maxhaider.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^3.2.8"
django-rest-framework = "^0.1.0"
django-extensions = "^3.1.3"
django-debug-toolbar = "^3.2.2"
psycopg2-binary = "^2.9.9"
gunicorn = "^20.1.0"
whitenoise = "^6.7.0"
factory-boy = "^3.3.1"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
